+++ packages/ub-pub/CHANGELOG.md
@@ -15,10 +15,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - added `filter i18n` to UB.install. Can be used as $ut replacement
 ```vuejs
 <div> {{ 'uba_user' | i18n}} </div 
+<div> {{ 'login' | i18n('Mike') }} </div
 ```
 is equal to
 ```js
 <div> {{ $ut('uba_user') }} </div 
+<div> {{ $ut(login, 'Mike') }} </div
 ```
   
 ## [5.3.5]